+++ YDS-Sample/AtomSampleViewController/SimpleTextFieldSampleViewController.swift
@@ -11,8 +11,8 @@ import SnapKit
 
 class SimpleTextFieldSampleViewController: UIViewController, UITextFieldDelegate {
     
-    let sampleTextField: YDSSimpleTextField = {
-        let textField = YDSSimpleTextField()
+    let sampleTextField: YDSSimpleTextFieldView = {
+        let textField = YDSSimpleTextFieldView()
         textField.fieldLabelText = "닉네임"
         textField.placeholder = "John Doe"
         textField.helperLabelText = "이상한거 쓰지 마세요."
@@ -37,14 +37,14 @@ class SimpleTextFieldSampleViewController: UIViewController, UITextFieldDelegate
             make.trailing.equalToSuperview().offset(-20)
         }
         
-        _ = sampleTextField.base.becomeFirstResponder()
-        sampleTextField.base.delegate = self
-        sampleTextField.base.returnKeyType = .done
-        sampleTextField.base.addTarget(self, action: #selector(self.textFieldDidChange(_:)), for: .editingChanged)
+        _ = sampleTextField.textField.becomeFirstResponder()
+        sampleTextField.textField.delegate = self
+        sampleTextField.textField.returnKeyType = .done
+        sampleTextField.textField.addTarget(self, action: #selector(self.textFieldDidChange(_:)), for: .editingChanged)
     }
     
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        sampleTextField.base.endEditing(true)
+        sampleTextField.textField.endEditing(true)
     }
     
     @objc func textFieldDidChange(_ textField: UITextField) {
